Struct blpapi::session::SessionSync

pub struct SessionSync(_);

A wrapper for session which only show sync fn

Methods

impl SessionSync

pub fn from_options(options: SessionOptions) -> Self

Create a new SessionSync from a SessionOptions

pub fn new() -> Result<Self, Error>

Create a new SessionSync with default options and open refdata service

pub fn send(
&mut self,
request: Request,
correlation_id: Option<CorrelationId>
) -> Result<Events, Error>

Send request and get Events iterator

pub fn next_event(&mut self, timeout_ms: Option<u32>) -> Result<Event, Error>

Request for next event, optionally waiting timeout_ms if there is no event

pub fn ref_data<I, R>(
&mut self,
securities: I
) -> Result<HashMap<String, R>, Error> where
I: IntoIterator,
I::Item: AsRef<str>,
R: RefData,

Get reference data for RefData items

Note

For ease of use, you can activate the derive feature.

Example

use blpapi::{RefData, session::SessionSync};

// use the **derive** feature to automatically convert field names into bloomberg fields
#[derive(Default, RefData)]
struct EquityData {
    ticker: String,
    crncy: String,
    market_status: Option<String>,
}

let mut session = SessionSync::new().unwrap();
let securities: &[&str] = &[ /* list of security tickers */ ];

let maybe_equities = session.ref_data::<_, EquityData>(securities);

pub fn hist_data<I, R>(
&mut self,
securities: I,
options: HistOptions
) -> Result<HashMap<String, TimeSerie<R>>, Error> where
I: IntoIterator,
I::Item: AsRef<str>,
R: RefData,

Get historical data

Note

For ease of use, you can activate the derive feature.

Example

use blpapi::{RefData, session::{SessionSync, HistOptions}};

// use the **derive** feature to automatically convert field names into bloomberg fields
#[derive(Default, RefData)]
struct Price {
    px_last: f64,
}

let mut session = SessionSync::new().unwrap();
let securities: &[&str] = &[ /* list of security tickers */ ];

let options = HistOptions::new("20190101", "20191231");
let prices = session.hist_data::<_, Price>(securities, options);